backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.3
PyMySQL==1.1.0
Flask-Migrate==4.0.4
Flask-JWT-Extended==4.7.1
python-dotenv==1.0.0
passlib[bcrypt]==1.7.4
Flask-Cors==4.0.0
